Time: 0 ms (100%), Space: 22.7 MB (50.89%) - LeetHub

+++ 0026-remove-duplicates-from-sorted-array/0026-remove-duplicates-from-sorted-array.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     int removeDuplicates(vector<int>& nums) {
         if (nums.empty()) return 0;
-        int j = 0;
-        for (int i = 1; i < nums.size(); i++)
+        unsigned short int j = 0;
+        for (unsigned short int i = 1; i < nums.size(); i++)
         {
             //j is droped by till we find different
             if (nums[i] != nums[j])
